Move SwordSkillInventory specific save tasks
